+++ kit/lib/apollo.js
@@ -7,7 +7,12 @@
 import { createNetworkInterface, ApolloClient } from 'react-apollo';
 
 /* ReactQL */
+
+// Configuration
 import config from 'kit/config';
+
+// Get environment, to figure out where we're running the GraphQL server
+import { getServerURL } from 'kit/lib/env';
 
 // ----------------------
 
@@ -21,9 +26,16 @@ export function createClient(opt = {}) {
 
 // Creates a new browser client
 export function browserClient() {
+  // If we have an internal GraphQL server, we need to append it with a
+  // call to `getServerURL()` to add the correct host (in dev + production)
+  const uri = config.graphQLServer
+    ? `${getServerURL()}${config.graphQLEndpoint}` : config.graphQLEndpoint;
+
   return createClient({
     networkInterface: createNetworkInterface({
-      uri: config.graphQLEndpoint,
+      // If we have an internal GraphQL server, then we should append the
+      // URL with `getServerURL()` to get the correct hostname
+      uri,
     }),
   });
 }